import build fixes from 1.4.2
git-svn-id: svn://scm.orgis.org/mpg123/trunk@1472 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libmpg123/readers.c
@@ -124,7 +124,7 @@ static ssize_t icy_fullread(mpg123_handle *fr, unsigned char *buf, ssize_t count
 				fr->icy.next -= ret;
 				if(fr->icy.next > 0)
 				{
-					debug1("another try... still %li left", fr->icy.next);
+					debug1("another try... still %li left", (long)fr->icy.next);
 					continue;
 				}
 			}
